sexpr literals in any radix

+++ src/main/scala/regolic/sexpr/SExprs.scala
@@ -7,10 +7,10 @@ object SExprs {
   object SList {
     def apply(sexprs: SExpr*): SList = SList(List(sexprs:_*))
   }
-  case class SInt(n: Int) extends SExpr
+  case class SInt(n: BigInt) extends SExpr
   case class SDouble(n: Double) extends SExpr
   case class SString(s: String) extends SExpr
   case class SSymbol(s: String) extends SExpr
-  case class SComment(s: String) extends SExpr
+  case class SComment(s: String) extends SExpr /* Never parsed, only used for pretty printing */
 
 }
