+++ src/main/scala/regolic/sexpr/Tokens.scala
@@ -9,7 +9,7 @@ object Tokens {
 
   case class StringLit(s: String) extends Token /* "hello" */
   case class SymbolLit(s: String) extends Token /* hello */
-  case class IntLit(n: Int)       extends Token /* 42 */
+  case class IntLit(n: BigInt)    extends Token /* 42, #b101, #xFF1D */
   case class DoubleLit(d: Double) extends Token /* 42.24 */
 
 }
